requirements.txt
# Core dependencies for Bhagavad Gita Chatbot
streamlit>=1.28.0
datasets>=2.14.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
# numpy 1.26+ has manylinux wheels for Python 3.13
numpy>=1.26.0,<2.0.0
# pandas 2.2+ provides wheels for Python 3.13; 1.x fails to build
pandas>=2.2.3,<3.0.0
scikit-learn>=1.3.0
torch>=1.11.0
transformers>=4.30.0

# Optional API integrations (app works without these)
google-generativeai>=0.3.0
openai>=1.3.0
python-dotenv>=1.0.0
